Fixes the over-logging of TextModified
We were overlogging the TextMoified action because the old
implementation would log a cursor change as a text
modification. This CL logs 4 specific actions (choose
suggestion, delete text, insert text, insert punctuation)
as a text modification rather than any action in the textbox

I also add in logging of the length of the recognition result
so that we can get some more context around the scope of the
editing of the ime result.

Change-Id: I172df24ddc0a7b62bcc5ed806fd70ef7e1c42310