Refactor popup keyboard handling
Fixes issue 9.

The original code had a special case for digits as the leftmost character
of the popup keyboard, and used that to decide if it wants to align the
left or right edge of the popup above the key being held.

The new logic always aligns the popup keyboard so that it extends towards
the center of the screen.

Also remove digits from popups for the qwerty row in full mode.

This change also updates the popup key mappings to match the new logic.
I had to remove a few symbols on "4" to fit, Lira/Peso/Franc are now on
the L/P/F keys.