Revive the slide-up extension keyboard from AOSP
The AOSP release contained support for showing additional key rows when sliding
up on the keyboard, providing digits and punctuation in the 4-row portrait
mode. This was apparently disabled and replaced with long-press alternates, but
I think it's useful and quicker than long-pressing. Re-enabling the code. At
this point it interferes with long-press popups to some extent, working on
fixing that separately.

This also adds a layout for a single extension row in 5-row mode.

I reverted the following git changelist as the starting point:

http://code.google.com/p/hackerskeyboard/source/detail?r=91606433019c0b0b2a49a4a7c85379a235d685ad#

    commit d2a431efa726771dee5c7b90004a0ed670d9a129
    Author: Tadashi G. Takaoka <takaoka@google.com>
    Date:   Mon Aug 30 21:09:44 2010 +0900

        Remove extension keyboard

        Bug: 2959287
        Change-Id: I344ac58b4119e96daf5928c784275ee0b7095958