Remove unused code for conditional settings key support
The AOSP code contains logic to conditionally hide the settings key when
LatinIME is the only enabled input method. This was reasonable for its use as
the default system IME, but the condition is always false for Hacker's Keyboard
since it's an add-on IME, and the system one can't be disabled.

In addition to simplifying things, this also avoids the extremely frequent
re-reading of the preferences XML file and system service lookups it was doing
while checking if the number of IMEs has changed.